+++ src/ServiceBus/ServiceBus.Autorest/custom/New-AzServiceBusNamespace.ps1
@@ -167,7 +167,21 @@ function New-AzServiceBusNamespace{
         [Microsoft.Azure.PowerShell.Cmdlets.ServiceBus.Category('Runtime')]
         [System.Management.Automation.SwitchParameter]
         # Use the default credentials for the proxy
-        ${ProxyUseDefaultCredentials}
+        ${ProxyUseDefaultCredentials},
+
+        [Parameter(HelpMessage = "The maximum acceptable lag for data replication operations from the primary replica to a quorum of secondary replicas. When the lag exceeds the configured amount, operations on the primary replica will be failed. The allowed values are 0 and 5 minutes to 1 day.")]
+		[Microsoft.Azure.PowerShell.Cmdlets.ServiceBus.Category('Body')]
+		[System.Int64]
+		${GeoDataReplicationMaxReplicationLagDurationInSecond},
+
+		[Parameter(HelpMessage = "Properties for User Assigned Identities")]
+		[Microsoft.Azure.PowerShell.Cmdlets.ServiceBus.Models.INamespaceReplicaLocation[]]
+		${GeoDataReplicationLocation},
+
+        [Parameter(HelpMessage = "The IP address type for the namespace. Determines whether the namespace supports IPv4 only or both IPv4 and IPv6.")]
+        [Microsoft.Azure.PowerShell.Cmdlets.ServiceBus.Category('Body')]
+        [System.String]
+        ${IPAddressType}
 	)
 	process{
 	    try{