+++ src/ServiceBus/ServiceBus.Autorest/custom/Set-AzServiceBusNamespace.ps1
@@ -159,7 +159,22 @@ function Set-AzServiceBusNamespace{
         [Microsoft.Azure.PowerShell.Cmdlets.ServiceBus.Category('Runtime')]
         [System.Management.Automation.SwitchParameter]
         # Use the default credentials for the proxy
-        ${ProxyUseDefaultCredentials}
+        ${ProxyUseDefaultCredentials},
+
+		[Parameter(HelpMessage = "The maximum acceptable lag for data replication operations from the primary replica to a quorum of secondary replicas. When the lag exceeds the configured amount, operations on the primary replica will be failed. The allowed values are 0 and 5 minutes to 1 day.")]
+		[Microsoft.Azure.PowerShell.Cmdlets.ServiceBus.Category('Body')]
+		[System.Int64]
+		${GeoDataReplicationMaxReplicationLagDurationInSecond},
+
+		[Parameter(HelpMessage = "Properties for User Assigned Identities")]
+		[Microsoft.Azure.PowerShell.Cmdlets.ServiceBus.Category('Body')]
+		[Microsoft.Azure.PowerShell.Cmdlets.ServiceBus.Models.INamespaceReplicaLocation[]]
+		${GeoDataReplicationLocation},
+
+        [Parameter(HelpMessage = "The IP address type for the namespace. Determines whether the namespace supports IPv4 only or both IPv4 and IPv6.")]
+        [Microsoft.Azure.PowerShell.Cmdlets.ServiceBus.Category('Body')]
+        [System.String]
+        ${IPAddressType}
 	)
 	process{
 	    try{
@@ -173,6 +188,9 @@ function Set-AzServiceBusNamespace{
                 $hasPublicNetworkAccess = $PSBoundParameters.Remove('PublicNetworkAccess')
                 $hasSkuCapacity = $PSBoundParameters.Remove('SkuCapacity')
                 $hasTag = $PSBoundParameters.Remove('Tag')
+                $hasGeoDataReplicationMaxReplicationLagDurationInSecond = $PSBoundParameters.Remove('GeoDataReplicationMaxReplicationLagDurationInSecond')
+                $hasGeoDataReplicationLocation = $PSBoundParameters.Remove('GeoDataReplicationLocation')
+                $hasIPAddressType = $PSBoundParameters.Remove('IPAddressType')
                 $hasDefaultProfile = $PSBoundParameters.Remove('DefaultProfile')
                 $hasAsJob = $PSBoundParameters.Remove('AsJob')
                 $hasNoWait = $PSBoundParameters.Remove('NoWait')
@@ -223,6 +241,15 @@ function Set-AzServiceBusNamespace{
                }
                if ($hasTag) {
                     $serviceBusNamespace.Tag = $Tag
+               }
+               if ($hasGeoDataReplicationMaxReplicationLagDurationInSecond) {
+                    $serviceBusNamespace.GeoDataReplicationMaxReplicationLagDurationInSecond = $GeoDataReplicationMaxReplicationLagDurationInSecond
+               }
+               if ($hasGeoDataReplicationLocation) {
+                    $serviceBusNamespace.GeoDataReplicationLocation = $GeoDataReplicationLocation
+               }
+               if($hasIPAddressType){
+                   $serviceBusNamespace.IPAddressType = $IPAddressType
                }
                if ($hasAsJob) {
                     $PSBoundParameters.Add('AsJob', $true)