+++ src/ServiceBus/ServiceBus/ChangeLog.md
@@ -19,6 +19,9 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added parameters 'MaxReplicationLagDurationInSeconds' and 'Replicalocation' and 'Ipaddresstype' in cmdlets 'New-AzServiceBusNamespace' & 'Set-AzServiceBusNamespace'
+* Added new cmdlets for doing namespace failover - 'Start-AzServiceBusNamespaceFailOver'
+* Added new cmdlets for generating replication object for Replication parameter - 'New-AzServiceBusLocationsNameObject'
 
 ## Version 4.2.0
 * Added ChangeSafety Support